Consolidate connection handling code
