+++ lib/qbo_api.rb
@@ -3,10 +3,8 @@ require 'json'
 require 'uri'
 require 'securerandom'
 require 'logger'
-require 'faraday'
-require 'faraday_middleware'
-require 'faraday/detailed_logger'
 require_relative 'qbo_api/configuration'
+require_relative 'qbo_api/connection'
 require_relative 'qbo_api/supporting'
 require_relative 'qbo_api/error'
 require_relative 'qbo_api/raise_http_exception'
@@ -20,6 +18,7 @@ require_relative 'qbo_api/all'
 
 class QboApi
   extend Configuration
+  include Connection
   include Supporting
   include Entity
   include Util
@@ -52,15 +51,7 @@ class QboApi
   end
 
   def connection(url: get_endpoint)
-    @connection ||= Faraday.new(url: url) do |faraday|
-      faraday.headers['Content-Type'] = 'application/json;charset=UTF-8'
-      faraday.headers['Accept'] = 'application/json'
-      add_authorization_middleware(faraday)
-      faraday.request :url_encoded
-      faraday.use FaradayMiddleware::RaiseHttpException
-      faraday.response :detailed_logger, QboApi.logger, LOG_TAG if QboApi.log
-      faraday.adapter  Faraday.default_adapter
-    end
+    @connection ||= authorized_json_connection(url: url)
   end
 
   def create(entity, payload:, params: nil)
@@ -98,56 +89,7 @@ class QboApi
     request(:get, path: path)
   end
 
-  def request(method, path:, entity: nil, payload: nil, params: nil)
-    raw_response = connection.send(method) do |req|
-      path = finalize_path(path, method: method, params: params)
-      case method
-      when :get, :delete
-        req.url path
-      when :post, :put
-        req.url path
-        req.body = payload.to_json
-      end
-    end
-    response(raw_response, entity: entity)
-  end
-
-  def response(resp, entity: nil)
-    data = parse_response_body(resp)
-    entity ? entity_response(data, entity) : data
-  rescue => e
-    QboApi.logger.debug { "#{LOG_TAG} response parsing error: entity=#{entity.inspect} body=#{resp.body.inspect} exception=#{e.inspect}" }
-    data
-  end
-
-  def parse_response_body(resp)
-    body = resp.body
-    case resp.headers['Content-Type']
-    when /json/ then JSON.parse(body)
-    else body
-    end
-  end
-
   private
-
-  def entity_response(data, entity)
-    entity_name = entity_name(entity)
-    if data.key?('QueryResponse')
-      entity_body = data['QueryResponse']
-      return nil if entity_body.empty?
-      entity_body.fetch(entity_name, data)
-    elsif data.key?('AttachableResponse')
-      entity_body = data['AttachableResponse']
-      entity_body &&= entity_body.first
-      entity_body.fetch(entity_name, data)
-    else
-      entity_body = data
-      entity_body.fetch(entity_name) do
-        QboApi.logger.debug { "#{LOG_TAG} entity name not in response body: entity=#{entity.inspect} entity_name=#{entity_name.inspect} body=#{data.inspect}" }
-        data
-      end
-    end
-  end
 
   def get_endpoint
     prod = self.class.production
@@ -157,32 +99,5 @@ class QboApi
     when :payments
       prod ? PAYMENTS_API_BASE_URL.sub("sandbox.", '') : PAYMENTS_API_BASE_URL
     end
-  end
-
-  def add_authorization_middleware(faraday)
-    if @token != nil
-      gem 'simple_oauth'
-      require 'simple_oauth'
-      faraday.request :oauth, oauth_data
-    elsif @access_token != nil
-      faraday.request :oauth2, @access_token, token_type: 'bearer'
-    else
-      raise QboApi::Error.new error_body: 'Must set either the token or access_token'
-    end
-  end
-
-  # Use with simple_oauth OAuth1 middleware
-  # @see #add_authorization_middleware
-  def oauth_data
-    {
-      consumer_key: @consumer_key,
-      consumer_secret: @consumer_secret,
-      token: @token,
-      token_secret: @token_secret
-    }
-  end
-
-  def entity_name(entity)
-    singular(entity)
   end
 end
